+++ src/app/_models/chart.ts
@@ -1,6 +1,5 @@
 import {DataSource} from './data.source';
 
-export const ChartGroupBy = ['10m', '1h', '1d'];
 export const ChartTypes = ['Bar', 'Line', 'Wind Arrow'];
 export type ChartType = 'Bar' | 'Line' | 'Wind Arrow';
 
@@ -13,7 +12,6 @@ export class Chart {
     color: string;
     decimals: number;
     type: ChartType;
-    group_by: string;
     query: string;
     data_source_id: number;
     data_source: DataSource;